# atch/src/main/java/com/jakaysa/atch/infrastructure/adapters/in/rest/dto/ErrorResponse.java
package com.jakaysa.atch.infrastructure.adapters.in.rest.dto;

import lombok.AllArgsConstructor;
import lombok.Data;

@Data
@AllArgsConstructor
public class ErrorResponse {
    private String error;
}